feat: add Dockerfile
