+++ package.json
@@ -66,7 +66,7 @@
     "eslint-plugin-import": "^2.27.5",
     "jest": "29.5.0",
     "prettier": "^2.8.8",
-    "rimraf": "^5.0.9",
+    "rimraf": "^5.0.10",
     "supertest": "^6.3.3",
     "ts-jest": "29.1.0",
     "ts-loader": "^9.4.3",
